Moved my SignInSignUp code to a separate app.  This is much cleaner.
