+++ django_project/settings.py
@@ -42,6 +42,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'RideShare',
+    'SignUpSignIn',
 ]
 
 MIDDLEWARE = [
