+++ django_project/urls.py
@@ -19,4 +19,5 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('RideShare.urls')),
+    path('', include('SignUpSignIn.urls')),
 ]
